+++ server/impl/src/main/java/com/walmartlabs/concord/server/cfg/LdapConfiguration.java
@@ -10,7 +10,10 @@ import java.io.InputStream;
 import java.io.Serializable;
 import java.nio.file.Files;
 import java.nio.file.Paths;
+import java.util.Collections;
+import java.util.HashSet;
 import java.util.Properties;
+import java.util.Set;
 
 @Named
 @Singleton
@@ -25,6 +28,7 @@ public class LdapConfiguration implements Serializable {
     private final String principalSearchFilter;
     private final String systemUsername;
     private final String systemPassword;
+    private final Set<String> exposeAttributes;
 
     public LdapConfiguration() throws IOException {
         Properties props = new Properties();
@@ -48,15 +52,7 @@ public class LdapConfiguration implements Serializable {
         this.principalSearchFilter = props.getProperty("principalSearchFilter");
         this.systemUsername = props.getProperty("systemUsername");
         this.systemPassword = props.getProperty("systemPassword");
-    }
-
-    public LdapConfiguration(String url, String searchBase, String principalSuffix, String principalSearchFilter, String systemUsername, String systemPassword) {
-        this.url = url;
-        this.searchBase = searchBase;
-        this.principalSuffix = principalSuffix;
-        this.principalSearchFilter = principalSearchFilter;
-        this.systemUsername = systemUsername;
-        this.systemPassword = systemPassword;
+        this.exposeAttributes = split(props, "exposeAttributes");
     }
 
     public String getUrl() {
@@ -81,5 +77,29 @@ public class LdapConfiguration implements Serializable {
 
     public String getSystemPassword() {
         return systemPassword;
+    }
+
+    public Set<String> getExposeAttributes() {
+        return exposeAttributes;
+    }
+
+    private static Set<String> split(Properties props, String key) {
+        String s = props.getProperty(key);
+        if (s == null || s.isEmpty()) {
+            return Collections.emptySet();
+        }
+
+        s = s.trim();
+        if (s.isEmpty()) {
+            return Collections.emptySet();
+        }
+
+        String[] as = s.split(",");
+        Set<String> result = new HashSet<>(as.length);
+        for (String item : s.split(",")) {
+            result.add(item);
+        }
+
+        return Collections.unmodifiableSet(result);
     }
 }