+++ server/impl/src/main/java/com/walmartlabs/concord/server/console/ConsoleService.java
@@ -1,5 +1,6 @@
 package com.walmartlabs.concord.server.console;
 
+import com.walmartlabs.concord.server.api.process.ProcessResource;
 import com.walmartlabs.concord.server.security.UserPrincipal;
 import com.walmartlabs.concord.server.security.ldap.LdapInfo;
 import org.apache.shiro.SecurityUtils;
@@ -7,6 +8,7 @@ import org.apache.shiro.authz.annotation.RequiresAuthentication;
 import org.apache.shiro.subject.Subject;
 import org.sonatype.siesta.Resource;
 
+import javax.inject.Inject;
 import javax.inject.Named;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
@@ -15,6 +17,13 @@ import javax.ws.rs.core.Response;
 @Named
 @Path("/api/service/console")
 public class ConsoleService implements Resource {
+
+    private final ProcessResource processResource;
+
+    @Inject
+    public ConsoleService(ProcessResource processResource) {
+        this.processResource = processResource;
+    }
 
     @GET
     @Path("/whoami")