+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/pipelines/ArchivePipeline.java
@@ -5,7 +5,6 @@ import com.walmartlabs.concord.server.ansible.InlineInventoryProcessor;
 import com.walmartlabs.concord.server.ansible.InventoryProcessor;
 import com.walmartlabs.concord.server.ansible.PrivateKeyProcessor;
 import com.walmartlabs.concord.server.process.pipelines.processors.*;
-import com.walmartlabs.concord.server.project.RepositoryProcessor;
 import com.walmartlabs.concord.server.template.TemplateProcessor;
 
 import javax.inject.Inject;
@@ -33,6 +32,7 @@ public class ArchivePipeline extends Chain {
                 injector.getInstance(PrivateKeyProcessor.class),
                 injector.getInstance(TemplateProcessor.class),
                 injector.getInstance(DependenciesProcessor.class),
+                injector.getInstance(UserInfoProcessor.class),
                 injector.getInstance(RequestDataStoringProcessor.class),
                 injector.getInstance(RunnerProcessor.class),
                 injector.getInstance(LogFileProcessor.class),