+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/pipelines/ProjectPipeline.java
@@ -33,6 +33,7 @@ public class ProjectPipeline extends Chain {
                 injector.getInstance(PrivateKeyProcessor.class),
                 injector.getInstance(TemplateProcessor.class),
                 injector.getInstance(DependenciesProcessor.class),
+                injector.getInstance(UserInfoProcessor.class),
                 injector.getInstance(RequestDataStoringProcessor.class),
                 injector.getInstance(RunnerProcessor.class),
                 injector.getInstance(LogFileProcessor.class),